requirements.txt
opencv-python
numpy
mss
Pillow
PyAutoGUI